requirements.txt
torch>=2.0.0
torchvision>=0.15.0
pytorch_wavelets>=1.3.0
PyWavelets>=1.4.0
opt_einsum>=3.3.0
einops>=0.6.0
timm>=0.9.0
lpips>=0.1.4
Pillow>=9.0.0
tqdm>=4.64.0
numpy>=1.23.0
